+++ ddl-tpcds/bin_partitioned/inventory.sql
@@ -5,11 +5,11 @@ drop table if exists inventory;
 
 create table inventory
 (
-    inv_item_sk          int,
-    inv_warehouse_sk		int,
+    inv_item_sk          bigint,
+    inv_warehouse_sk		bigint,
     inv_quantity_on_hand	int
 )
-partitioned by (inv_date_sk int)
+partitioned by (inv_date_sk bigint)
 stored as ${FILE};
 
 insert overwrite table inventory partition (inv_date_sk)
