+++ ddl-tpcds/bin_partitioned/store_returns.sql
@@ -5,27 +5,27 @@ drop table if exists store_returns;
 
 create table store_returns
 (
-    sr_return_time_sk         int,
-    sr_item_sk                int,
-    sr_customer_sk            int,
-    sr_cdemo_sk               int,
-    sr_hdemo_sk               int,
-    sr_addr_sk                int,
-    sr_store_sk               int,
-    sr_reason_sk              int,
+    sr_return_time_sk         bigint,
+    sr_item_sk                bigint,
+    sr_customer_sk            bigint,
+    sr_cdemo_sk               bigint,
+    sr_hdemo_sk               bigint,
+    sr_addr_sk                bigint,
+    sr_store_sk               bigint,
+    sr_reason_sk              bigint,
     sr_ticket_number          bigint,
     sr_return_quantity        int,
-    sr_return_amt             float,
-    sr_return_tax             float,
-    sr_return_amt_inc_tax     float,
-    sr_fee                    float,
-    sr_return_ship_cost       float,
-    sr_refunded_cash          float,
-    sr_reversed_charge        float,
-    sr_store_credit           float,
-    sr_net_loss               float
+    sr_return_amt             double,
+    sr_return_tax             double,
+    sr_return_amt_inc_tax     double,
+    sr_fee                    double,
+    sr_return_ship_cost       double,
+    sr_refunded_cash          double,
+    sr_reversed_charge        double,
+    sr_store_credit           double,
+    sr_net_loss               double
 )
-partitioned by (sr_returned_date_sk int)
+partitioned by (sr_returned_date_sk bigint)
 stored as ${FILE};
 
 insert overwrite table store_returns partition (sr_returned_date_sk)
