+++ ddl-tpcds/bin_partitioned/store_sales.sql
@@ -5,30 +5,30 @@ drop table if exists store_sales;
 
 create table store_sales
 (
-    ss_sold_time_sk           int,
-    ss_item_sk                int,
-    ss_customer_sk            int,
-    ss_cdemo_sk               int,
-    ss_hdemo_sk               int,
-    ss_addr_sk                int,
-    ss_store_sk               int,
-    ss_promo_sk               int,
+    ss_sold_time_sk           bigint,
+    ss_item_sk                bigint,
+    ss_customer_sk            bigint,
+    ss_cdemo_sk               bigint,
+    ss_hdemo_sk               bigint,
+    ss_addr_sk                bigint,
+    ss_store_sk               bigint,
+    ss_promo_sk               bigint,
     ss_ticket_number          bigint,
     ss_quantity               int,
-    ss_wholesale_cost         float,
-    ss_list_price             float,
-    ss_sales_price            float,
-    ss_ext_discount_amt       float,
-    ss_ext_sales_price        float,
-    ss_ext_wholesale_cost     float,
-    ss_ext_list_price         float,
-    ss_ext_tax                float,
-    ss_coupon_amt             float,
-    ss_net_paid               float,
-    ss_net_paid_inc_tax       float,
-    ss_net_profit             float
+    ss_wholesale_cost         double,
+    ss_list_price             double,
+    ss_sales_price            double,
+    ss_ext_discount_amt       double,
+    ss_ext_sales_price        double,
+    ss_ext_wholesale_cost     double,
+    ss_ext_list_price         double,
+    ss_ext_tax                double,
+    ss_coupon_amt             double,
+    ss_net_paid               double,
+    ss_net_paid_inc_tax       double,
+    ss_net_profit             double
 )
-partitioned by (ss_sold_date_sk int)
+partitioned by (ss_sold_date_sk bigint)
 stored as ${FILE};
 
 insert overwrite table store_sales partition (ss_sold_date_sk)
